refactor: extract doctor commands into separate modules
- Move each command implementation to its own file in src/commands/
- Export necessary types and functions from doctor.rs for command modules
- Clean up imports and remove duplicated code
- No functional changes, just code organization
