+++ src/commands/mod.rs
@@ -1,0 +1,8 @@
+//! Doctor command implementations
+
+pub mod accept;
+pub mod connect;
+pub mod plot;
+pub mod port_map;
+pub mod relay_urls;
+pub mod report;
