+++ src/commands/port_map.rs
@@ -1,0 +1,56 @@
+//! Port map command implementation
+
+use std::{num::NonZeroU16, time::Duration};
+
+use n0_watcher::Watcher;
+
+/// Checks if there's a port mapping in the local port, and if it's ready.
+pub async fn port_map(
+    protocol: &str,
+    local_port: NonZeroU16,
+    timeout: Duration,
+) -> anyhow::Result<()> {
+    // Create the config that enables exclusively the required protocol
+    let mut enable_upnp = false;
+    let mut enable_pcp = false;
+    let mut enable_nat_pmp = false;
+    match protocol.to_ascii_lowercase().as_ref() {
+        "upnp" => enable_upnp = true,
+        "nat_pmp" => enable_nat_pmp = true,
+        "pcp" => enable_pcp = true,
+        other => anyhow::bail!("Unknown port mapping protocol {other}"),
+    }
+    let config = portmapper::Config {
+        enable_upnp,
+        enable_pcp,
+        enable_nat_pmp,
+    };
+    let port_mapper = portmapper::Client::new(config);
+    let mut watcher = port_mapper.watch_external_address();
+    port_mapper.update_local_port(local_port);
+
+    // Wait for the mapping to be ready, or timeout waiting for a change.
+    match tokio::time::timeout(timeout, watcher.changed()).await {
+        Ok(Ok(_)) => match *watcher.borrow() {
+            Some(address) => {
+                println!("Port mapping ready: {address}");
+                // Ensure the port mapper remains alive until the end.
+                drop(port_mapper);
+                Ok(())
+            }
+            None => anyhow::bail!("No port mapping found"),
+        },
+        Ok(Err(_recv_err)) => anyhow::bail!("Service dropped. This is a bug"),
+        Err(_) => anyhow::bail!("Timed out waiting for a port mapping"),
+    }
+}
+
+/// Probes a port map.
+pub async fn port_map_probe(config: portmapper::Config) -> anyhow::Result<()> {
+    println!("probing port mapping protocols with {config:?}");
+    let port_mapper = portmapper::Client::new(config);
+    let probe_rx = port_mapper.probe();
+    let probe = probe_rx.await?.map_err(|e| anyhow::anyhow!(e))?;
+    println!("{probe}");
+    Ok(())
+}
