+++ src/doctor.rs
@@ -2,47 +2,31 @@
 //! and to test connectivity to specific other nodes.
 
 use std::{
-    collections::HashMap,
-    io,
     net::SocketAddr,
     num::NonZeroU16,
     path::PathBuf,
-    sync::Arc,
     time::{Duration, Instant},
 };
 
 use anyhow::Context;
 use clap::Subcommand;
-use crossterm::{
-    event::{self, DisableMouseCapture, EnableMouseCapture, Event, KeyCode, KeyEventKind},
-    execute,
-    terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
-};
-use futures_lite::StreamExt;
-use futures_util::SinkExt;
 use indicatif::{HumanBytes, MultiProgress, ProgressBar};
 use iroh::{
     discovery::{dns::DnsDiscovery, pkarr::PkarrPublisher, ConcurrentDiscovery},
-    dns::DnsResolver,
     endpoint::{self, Connection, ConnectionType, RecvStream, RemoteInfo, SendStream},
     metrics::MagicsockMetrics,
-    Endpoint, NodeAddr, NodeId, RelayMap, RelayMode, RelayNode, RelayUrl, SecretKey, Watcher,
+    Endpoint, NodeId, RelayMap, RelayMode, RelayNode, RelayUrl, SecretKey,
 };
 use iroh_metrics::static_core::Core;
-use iroh_relay::{
-    protos::relay::{ClientToRelayMsg, RelayToClientMsg},
-    RelayQuicConfig,
-};
-use portable_atomic::AtomicU64;
+use n0_watcher::Watcher;
+use iroh_relay::RelayQuicConfig;
 use postcard::experimental::max_size::MaxSize;
-use rand::Rng;
-use ratatui::{prelude::*, widgets::*};
 use serde::{Deserialize, Serialize};
 use tokio::{io::AsyncWriteExt, sync};
 use tokio_util::task::AbortOnDropHandle;
-use tracing::warn;
 
 use crate::{
+    commands,
     config::{iroh_data_root, NodeConfig},
     metrics::{IrohMetricsRegistry, MetricsRegistry},
     progress::ProgressWriter,
@@ -231,7 +215,7 @@ pub enum Commands {
 
 /// Possible streams that can be requested.
 #[derive(Debug, Serialize, Deserialize, MaxSize)]
-enum TestStreamRequest {
+pub enum TestStreamRequest {
     Echo { bytes: u64 },
     Drain { bytes: u64 },
     Send { bytes: u64, block_size: u32 },
@@ -330,72 +314,24 @@ async fn send_blocks(
     Ok(())
 }
 
-/// Prints a client report.
-#[allow(clippy::too_many_arguments)]
-async fn report(
-    config: &NodeConfig,
-    mut quic_ipv4: bool,
-    mut quic_ipv6: bool,
-    mut https: bool,
-) -> anyhow::Result<()> {
-    // if all protocol flags are false, set them all to true
-    if !(quic_ipv4 || quic_ipv6 || https) {
-        quic_ipv4 = true;
-        quic_ipv6 = true;
-        https = true;
-    }
-    println!("Probe protocols selected:");
-    if quic_ipv4 {
-        println!("quic ipv4")
-    }
-    if quic_ipv6 {
-        println!("quic ipv6")
-    }
-    if https {
-        println!("https")
-    }
-    let relay_map = config.relay_map()?.unwrap_or_else(RelayMap::empty);
-
-    let endpoint = iroh::Endpoint::builder()
-        .relay_mode(RelayMode::Custom(relay_map.clone()))
-        .bind()
-        .await?;
-
-    println!("\nRelay Map:");
-    for (url, node) in relay_map.urls().zip(relay_map.nodes()) {
-        println!(
-            r#"- {url}
-  QUIC port: {:?}"#,
-            node.quic.as_ref().map(|c| c.port),
-        );
-    }
-
-    let mut stream = endpoint.net_report().stream();
-    while let Some(report) = stream.next().await {
-        println!("{report:#?}");
-    }
-
-    endpoint.close().await;
-    Ok(())
-}
 
 /// Contains all the GUI state.
-struct Gui {
+pub struct Gui {
     #[allow(dead_code)]
-    mp: MultiProgress,
-    pb: ProgressBar,
+    pub mp: MultiProgress,
+    pub pb: ProgressBar,
     #[allow(dead_code)]
-    counters: ProgressBar,
-    send_pb: ProgressBar,
-    recv_pb: ProgressBar,
-    echo_pb: ProgressBar,
+    pub counters: ProgressBar,
+    pub send_pb: ProgressBar,
+    pub recv_pb: ProgressBar,
+    pub echo_pb: ProgressBar,
     #[allow(dead_code)]
-    counter_task: Option<AbortOnDropHandle<()>>,
+    pub counter_task: Option<AbortOnDropHandle<()>>,
 }
 
 impl Gui {
     /// Create a new GUI struct.
-    fn new(endpoint: Endpoint, node_id: NodeId) -> Self {
+    pub fn new(endpoint: Endpoint, node_id: NodeId) -> Self {
         let mp = MultiProgress::new();
         mp.set_draw_target(indicatif::ProgressDrawTarget::stderr());
         let counters = mp.add(ProgressBar::hidden());
@@ -522,13 +458,13 @@ Ipv6:
     }
 
     /// Clears the [`MultiProgress`] field.
-    fn clear(&self) {
+    pub fn clear(&self) {
         self.mp.clear().ok();
     }
 }
 
 /// Sends, receives and echoes data in a connection.
-async fn active_side(
+pub async fn active_side(
     connection: &Connection,
     config: &TestConfig,
     gui: Option<&Gui>,
@@ -649,7 +585,7 @@ async fn recv_test(
 }
 
 /// Accepts connections and answers requests (echo, drain or send) as passive side.
-async fn passive_side(gui: Gui, connection: &Connection) -> anyhow::Result<()> {
+pub async fn passive_side(gui: Gui, connection: &Connection) -> anyhow::Result<()> {
     let conn = connection.clone();
     let accept_loop = async move {
         let result = loop {
@@ -767,44 +703,8 @@ async fn make_endpoint(
     Ok((endpoint, rpc_client))
 }
 
-/// Connects to a [`NodeId`].
-async fn connect(
-    node_id: NodeId,
-    direct_addresses: Vec<SocketAddr>,
-    relay_url: Option<RelayUrl>,
-    endpoint: Endpoint,
-) -> anyhow::Result<()> {
-    tracing::info!("dialing {:?}", node_id);
-    let node_addr = NodeAddr::from_parts(node_id, relay_url, direct_addresses);
-    let conn = endpoint.connect(node_addr, &DR_RELAY_ALPN).await;
-    match conn {
-        Ok(connection) => {
-            let maybe_conn_type = endpoint.conn_type(node_id);
-            let gui = Gui::new(endpoint, node_id);
-            if let Some(conn_type) = maybe_conn_type {
-                log_connection_changes(gui.mp.clone(), node_id, conn_type);
-            }
 
-            let close_reason = connection
-                .close_reason()
-                .map(|e| format!(" (reason: {e})"))
-                .unwrap_or_default();
-
-            if let Err(cause) = passive_side(gui, &connection).await {
-                eprintln!("error handling connection: {cause}{close_reason}");
-            } else {
-                eprintln!("Connection closed{close_reason}");
-            }
-        }
-        Err(cause) => {
-            eprintln!("unable to connect to {node_id}: {cause}");
-        }
-    }
-
-    Ok(())
-}
-
-async fn close_endpoint_on_ctrl_c(endpoint: Endpoint) {
+pub async fn close_endpoint_on_ctrl_c(endpoint: Endpoint) {
     tokio::signal::ctrl_c()
         .await
         .expect("failed listening to SIGINT");
@@ -812,7 +712,7 @@ async fn close_endpoint_on_ctrl_c(endpoint: Endpoint) {
 }
 
 /// Formats a [`SocketAddr`] so that console doesn't escape it.
-fn format_addr(addr: SocketAddr) -> String {
+pub fn format_addr(addr: SocketAddr) -> String {
     if addr.is_ipv6() {
         format!("'{addr}'")
     } else {
@@ -820,95 +720,9 @@ fn format_addr(addr: SocketAddr) -> String {
     }
 }
 
-/// Accepts the connections.
-pub async fn accept(
-    secret_key: SecretKey,
-    config: TestConfig,
-    endpoint: Endpoint,
-) -> anyhow::Result<()> {
-    let endpoints = endpoint.direct_addresses().initialized().await;
-
-    let remote_addrs = endpoints
-        .iter()
-        .map(|endpoint| format!("--remote-endpoint {}", format_addr(endpoint.addr)))
-        .collect::<Vec<_>>()
-        .join(" ");
-    println!("Connect to this node using one of the following commands:\n");
-    println!(
-        "\tUsing the relay url and direct connections:\niroh-doctor connect {} {}\n",
-        secret_key.public(),
-        remote_addrs,
-    );
-    if let Some(relay_url) = endpoint.home_relay().get().pop() {
-        println!(
-            "\tUsing just the relay url:\niroh-doctor connect {} --relay-url {}\n",
-            secret_key.public(),
-            relay_url,
-        );
-    }
-    if endpoint.discovery().is_some() {
-        println!(
-            "\tUsing just the node id:\niroh-doctor connect {}\n",
-            secret_key.public(),
-        );
-    }
-    let connections = Arc::new(AtomicU64::default());
-    while let Some(incoming) = endpoint.accept().await {
-        let connecting = match incoming.accept() {
-            Ok(connecting) => connecting,
-            Err(err) => {
-                warn!("incoming connection failed: {err:#}");
-                // we can carry on in these cases:
-                // this can be caused by retransmitted datagrams
-                continue;
-            }
-        };
-        let connections = connections.clone();
-        let endpoint = endpoint.clone();
-        tokio::task::spawn(async move {
-            let n = connections.fetch_add(1, portable_atomic::Ordering::SeqCst);
-            match connecting.await {
-                Ok(connection) => {
-                    if n == 0 {
-                        let Ok(remote_peer_id) = connection.remote_node_id() else {
-                            return;
-                        };
-                        println!("Accepted connection from {remote_peer_id}");
-                        let t0 = Instant::now();
-                        let gui = Gui::new(endpoint.clone(), remote_peer_id);
-                        if let Some(conn_type) = endpoint.conn_type(remote_peer_id) {
-                            log_connection_changes(gui.mp.clone(), remote_peer_id, conn_type);
-                        }
-                        let res = active_side(&connection, &config, Some(&gui)).await;
-                        gui.clear();
-                        let dt = t0.elapsed().as_secs_f64();
-                        if let Err(cause) = res {
-                            let close_reason = connection
-                                .close_reason()
-                                .map(|e| format!(" (reason: {e})"))
-                                .unwrap_or_default();
-                            eprintln!("Test finished after {dt}s: {cause}{close_reason}",);
-                        } else {
-                            eprintln!("Test finished after {dt}s",);
-                        }
-                    } else {
-                        // silent
-                        active_side(&connection, &config, None).await.ok();
-                    }
-                }
-                Err(cause) => {
-                    eprintln!("error accepting connection {cause}");
-                }
-            };
-            connections.sub(1, portable_atomic::Ordering::SeqCst);
-        });
-    }
-
-    Ok(())
-}
 
 /// Logs the connection changes to the multiprogress.
-fn log_connection_changes(
+pub fn log_connection_changes(
     pb: MultiProgress,
     node_id: NodeId,
     mut conn_type: n0_watcher::Direct<ConnectionType>,
@@ -925,193 +739,7 @@ fn log_connection_changes(
     });
 }
 
-/// Checks if there's a port mapping in the local port, and if it's ready.
-async fn port_map(protocol: &str, local_port: NonZeroU16, timeout: Duration) -> anyhow::Result<()> {
-    // Create the config that enables exclusively the required protocol
-    let mut enable_upnp = false;
-    let mut enable_pcp = false;
-    let mut enable_nat_pmp = false;
-    match protocol.to_ascii_lowercase().as_ref() {
-        "upnp" => enable_upnp = true,
-        "nat_pmp" => enable_nat_pmp = true,
-        "pcp" => enable_pcp = true,
-        other => anyhow::bail!("Unknown port mapping protocol {other}"),
-    }
-    let config = portmapper::Config {
-        enable_upnp,
-        enable_pcp,
-        enable_nat_pmp,
-    };
-    let port_mapper = portmapper::Client::new(config);
-    let mut watcher = port_mapper.watch_external_address();
-    port_mapper.update_local_port(local_port);
 
-    // Wait for the mapping to be ready, or timeout waiting for a change.
-    match tokio::time::timeout(timeout, watcher.changed()).await {
-        Ok(Ok(_)) => match *watcher.borrow() {
-            Some(address) => {
-                println!("Port mapping ready: {address}");
-                // Ensure the port mapper remains alive until the end.
-                drop(port_mapper);
-                Ok(())
-            }
-            None => anyhow::bail!("No port mapping found"),
-        },
-        Ok(Err(_recv_err)) => anyhow::bail!("Service dropped. This is a bug"),
-        Err(_) => anyhow::bail!("Timed out waiting for a port mapping"),
-    }
-}
-
-/// Probes a port map.
-async fn port_map_probe(config: portmapper::Config) -> anyhow::Result<()> {
-    println!("probing port mapping protocols with {config:?}");
-    let port_mapper = portmapper::Client::new(config);
-    let probe_rx = port_mapper.probe();
-    let probe = probe_rx.await?.map_err(|e| anyhow::anyhow!(e))?;
-    println!("{probe}");
-    Ok(())
-}
-
-/// Checks a certain amount (`count`) of the nodes given by the [`NodeConfig`].
-async fn relay_urls(count: usize, config: &NodeConfig) -> anyhow::Result<()> {
-    let key = SecretKey::generate(rand::rngs::OsRng);
-    if config.relay_nodes.is_empty() {
-        println!("No relay nodes specified in the config file.");
-    }
-
-    let dns_resolver = DnsResolver::new();
-    let mut client_builders = HashMap::new();
-    for node in &config.relay_nodes {
-        let secret_key = key.clone();
-        let client_builder = iroh_relay::client::ClientBuilder::new(
-            node.url.clone(),
-            secret_key,
-            dns_resolver.clone(),
-        );
-
-        client_builders.insert(node.url.clone(), client_builder);
-    }
-
-    let mut success = Vec::new();
-    let mut fail = Vec::new();
-
-    for i in 0..count {
-        println!("Round {}/{count}", i + 1);
-        let relay_nodes = config.relay_nodes.clone();
-        for node in relay_nodes.into_iter() {
-            let mut node_details = NodeDetails {
-                connect: None,
-                latency: None,
-                error: None,
-                host: node.url.clone(),
-            };
-
-            let client_builder = client_builders.get(&node.url).cloned().unwrap();
-
-            let start = std::time::Instant::now();
-            match tokio::time::timeout(Duration::from_secs(2), client_builder.connect()).await {
-                Err(e) => {
-                    tracing::warn!("connect timeout");
-                    node_details.error = Some(e.to_string());
-                }
-                Ok(Err(e)) => {
-                    tracing::warn!("connect error");
-                    node_details.error = Some(e.to_string());
-                }
-                Ok(Ok(client)) => {
-                    node_details.connect = Some(start.elapsed());
-                    match ping(client).await {
-                        Ok(latency) => {
-                            node_details.latency = Some(latency);
-                        }
-                        Err(e) => {
-                            tracing::warn!("ping error: {:?}", e);
-                            node_details.error = Some(e.to_string());
-                        }
-                    }
-                }
-            }
-
-            if node_details.error.is_none() {
-                success.push(node_details);
-            } else {
-                fail.push(node_details);
-            }
-        }
-    }
-
-    // success.sort_by_key(|d| d.latency);
-    if !success.is_empty() {
-        println!("Relay Node Latencies:");
-        println!();
-    }
-    for node in success {
-        println!("{node}");
-        println!();
-    }
-    if !fail.is_empty() {
-        println!("Connection Failures:");
-        println!();
-    }
-    for node in fail {
-        println!("{node}");
-        println!();
-    }
-
-    Ok(())
-}
-
-async fn ping(client: iroh_relay::client::Client) -> anyhow::Result<Duration> {
-    let (mut client_stream, mut client_sink) = client.split();
-    let data: [u8; 8] = rand::random();
-    let start = Instant::now();
-    client_sink.send(ClientToRelayMsg::Ping(data)).await?;
-    match tokio::time::timeout(Duration::from_secs(2), async move {
-        while let Some(res) = client_stream.next().await {
-            let res = res?;
-            if let RelayToClientMsg::Pong(d) = res {
-                if d == data {
-                    return Ok(start.elapsed());
-                }
-            }
-        }
-        anyhow::bail!("no pong received");
-    })
-    .await
-    {
-        Err(_) => {
-            anyhow::bail!("ping timeout");
-        }
-        Ok(res) => res,
-    }
-}
-
-/// Information about a node and its connection.
-struct NodeDetails {
-    connect: Option<Duration>,
-    latency: Option<Duration>,
-    host: RelayUrl,
-    error: Option<String>,
-}
-
-impl std::fmt::Display for NodeDetails {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self.error {
-            None => {
-                write!(
-                    f,
-                    "Node {}\nConnect: {:?}\nLatency: {:?}",
-                    self.host,
-                    self.connect.unwrap_or_default(),
-                    self.latency.unwrap_or_default(),
-                )
-            }
-            Some(ref err) => {
-                write!(f, "Node {}\nConnection Error: {:?}", self.host, err,)
-            }
-        }
-    }
-}
 
 /// Creates a [`SecretKey`] from a [`SecretKeyOption`].
 fn create_secret_key(secret_key: SecretKeyOption) -> anyhow::Result<SecretKey> {
@@ -1185,7 +813,7 @@ pub async fn run(
             quic_ipv4,
             quic_ipv6,
             https,
-        } => report(config, quic_ipv4, quic_ipv6, https).await,
+        } => commands::report::report(config, quic_ipv4, quic_ipv6, https).await,
         Commands::Connect {
             dial,
             secret_key,
@@ -1217,7 +845,7 @@ pub async fn run(
             .await?;
 
             futures_lite::future::race(close_endpoint_on_ctrl_c(endpoint.clone()), async move {
-                if let Err(e) = connect(dial, remote_endpoint, relay_url, endpoint).await {
+                if let Err(e) = commands::connect::connect(dial, remote_endpoint, relay_url, endpoint).await {
                     eprintln!("connect error: {e}");
                 }
             })
@@ -1254,7 +882,7 @@ pub async fn run(
             .await?;
 
             futures_lite::future::race(close_endpoint_on_ctrl_c(endpoint.clone()), async move {
-                if let Err(e) = accept(secret_key, config, endpoint).await {
+                if let Err(e) = commands::accept::accept(secret_key, config, endpoint).await {
                     eprintln!("accept error: {e}");
                 }
             })
@@ -1266,7 +894,7 @@ pub async fn run(
             protocol,
             local_port,
             timeout_secs,
-        } => port_map(&protocol, local_port, Duration::from_secs(timeout_secs)).await,
+        } => commands::port_map::port_map(&protocol, local_port, Duration::from_secs(timeout_secs)).await,
         Commands::PortMapProbe {
             enable_upnp,
             enable_pcp,
@@ -1278,9 +906,9 @@ pub async fn run(
                 enable_nat_pmp,
             };
 
-            port_map_probe(config).await
+            commands::port_map::port_map_probe(config).await
         }
-        Commands::RelayUrls { count } => relay_urls(count, config).await,
+        Commands::RelayUrls { count } => commands::relay_urls::relay_urls(count, config).await,
         Commands::Plot {
             interval,
             metrics,
@@ -1288,349 +916,13 @@ pub async fn run(
             scrape_url,
             file,
         } => {
-            let metrics: Vec<String> = metrics.split(',').map(|s| s.to_string()).collect();
-            let interval = Duration::from_millis(interval);
-
-            enable_raw_mode()?;
-            let mut stdout = io::stdout();
-            execute!(stdout, EnterAlternateScreen, EnableMouseCapture)?;
-            let backend = CrosstermBackend::new(stdout);
-            let mut terminal = Terminal::new(backend)?;
-
-            let app = PlotterApp::new(metrics, timeframe, scrape_url, file);
-            let res = run_plotter(&mut terminal, app, interval).await;
-            disable_raw_mode()?;
-            execute!(
-                terminal.backend_mut(),
-                LeaveAlternateScreen,
-                DisableMouseCapture
-            )?;
-            terminal.show_cursor()?;
-
-            if let Err(err) = res {
-                println!("{err:?}");
-            }
-
-            Ok(())
+            commands::plot::plot(interval, metrics, timeframe, scrape_url, file).await
         }
     };
     if let Some(metrics_fut) = metrics_fut {
         metrics_fut.abort();
     }
     cmd_res
-}
-
-/// Runs the [`PlotterApp`].
-async fn run_plotter<B: Backend>(
-    terminal: &mut Terminal<B>,
-    mut app: PlotterApp,
-    tick_rate: Duration,
-) -> anyhow::Result<()> {
-    let mut last_tick = Instant::now();
-    loop {
-        terminal.draw(|f| plotter_draw(f, &mut app))?;
-
-        if crossterm::event::poll(Duration::from_millis(10))? {
-            if let Event::Key(key) = event::read()? {
-                if key.kind == KeyEventKind::Press {
-                    if let KeyCode::Char(c) = key.code {
-                        app.on_key(c)
-                    }
-                }
-            }
-        }
-        if last_tick.elapsed() >= tick_rate {
-            app.on_tick().await;
-            last_tick = Instant::now();
-        }
-        if app.should_quit {
-            return Ok(());
-        }
-    }
-}
-
-/// Converts an area into `n` chunks.
-fn area_into_chunks(area: Rect, n: usize, is_horizontal: bool) -> std::rc::Rc<[Rect]> {
-    let mut constraints = vec![];
-    for _ in 0..n {
-        constraints.push(Constraint::Percentage(100 / n as u16));
-    }
-    let layout = match is_horizontal {
-        true => Layout::horizontal(constraints),
-        false => Layout::vertical(constraints),
-    };
-    layout.split(area)
-}
-
-/// Creates a collection of [`Rect`] by splitting an [`Rect`] area into `n` chunks.
-fn generate_layout_chunks(area: Rect, n: usize) -> Vec<Rect> {
-    if n < 4 {
-        let chunks = area_into_chunks(area, n, false);
-        return chunks.iter().copied().collect();
-    }
-    let main_chunks = area_into_chunks(area, 2, true);
-    let left_chunks = area_into_chunks(main_chunks[0], n / 2 + n % 2, false);
-    let right_chunks = area_into_chunks(main_chunks[1], n / 2, false);
-    let mut chunks = vec![];
-    chunks.extend(left_chunks.iter());
-    chunks.extend(right_chunks.iter());
-    chunks
-}
-
-/// Draws the [`Frame`] given a [`PlotterApp`].
-fn plotter_draw(f: &mut Frame, app: &mut PlotterApp) {
-    let area = f.area();
-
-    let metrics_cnt = app.metrics.len();
-    let areas = generate_layout_chunks(area, metrics_cnt);
-
-    for (i, metric) in app.metrics.iter().enumerate() {
-        plot_chart(f, areas[i], app, metric);
-    }
-}
-
-/// Draws the chart defined in the [`Frame`].
-fn plot_chart(frame: &mut Frame, area: Rect, app: &PlotterApp, metric: &str) {
-    let elapsed = app.internal_ts.as_secs_f64();
-    let data = app.data.get(metric).unwrap().clone();
-    let data_y_range = app.data_y_range.get(metric).unwrap();
-
-    let moved = (elapsed / 15.0).floor() * 15.0 - app.timeframe as f64;
-    let moved = moved.max(0.0);
-    let x_start = 0.0 + moved;
-    let x_end = moved + app.timeframe as f64 + 25.0;
-
-    let y_start = data_y_range.0;
-    let y_end = data_y_range.1;
-
-    let last_val = data.last();
-    let name = match last_val {
-        Some(val) => {
-            let val_y = val.1;
-            format!("{metric}: {val_y:.0}")
-        }
-        None => metric.to_string(),
-    };
-    let datasets = vec![Dataset::default()
-        .name(name)
-        .marker(symbols::Marker::Dot)
-        .graph_type(GraphType::Line)
-        .style(Style::default().fg(Color::Cyan))
-        .data(&data)];
-
-    // TODO(arqu): labels are incorrectly spaced for > 3 labels https://github.com/ratatui-org/ratatui/issues/334
-    let x_labels = vec![
-        Span::styled(
-            format!("{x_start:.1}s"),
-            Style::default().add_modifier(Modifier::BOLD),
-        ),
-        Span::raw(format!("{:.1}s", x_start + (x_end - x_start) / 2.0)),
-        Span::styled(
-            format!("{x_end:.1}s"),
-            Style::default().add_modifier(Modifier::BOLD),
-        ),
-    ];
-
-    let mut y_labels = vec![Span::styled(
-        format!("{y_start:.0}"),
-        Style::default().add_modifier(Modifier::BOLD),
-    )];
-
-    for i in 1..=10 {
-        y_labels.push(Span::raw(format!(
-            "{:.0}",
-            y_start + (y_end - y_start) / 10.0 * i as f64
-        )));
-    }
-
-    y_labels.push(Span::styled(
-        format!("{y_end:.0}"),
-        Style::default().add_modifier(Modifier::BOLD),
-    ));
-
-    let chart = Chart::new(datasets)
-        .block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title(format!("Chart: {metric}")),
-        )
-        .x_axis(
-            Axis::default()
-                .title("X Axis")
-                .style(Style::default().fg(Color::Gray))
-                .labels(x_labels)
-                .bounds([x_start, x_end]),
-        )
-        .y_axis(
-            Axis::default()
-                .title("Y Axis")
-                .style(Style::default().fg(Color::Gray))
-                .labels(y_labels)
-                .bounds([y_start, y_end]),
-        );
-
-    frame.render_widget(chart, area);
-}
-
-/// All the information about the plotter app.
-struct PlotterApp {
-    should_quit: bool,
-    metrics: Vec<String>,
-    start_ts: Instant,
-    data: HashMap<String, Vec<(f64, f64)>>,
-    data_y_range: HashMap<String, (f64, f64)>,
-    timeframe: usize,
-    rng: rand::rngs::ThreadRng,
-    freeze: bool,
-    internal_ts: Duration,
-    scrape_url: String,
-    file_data: Vec<String>,
-    file_header: Vec<String>,
-}
-
-impl PlotterApp {
-    /// Creates a new [`PlotterApp`].
-    fn new(
-        metrics: Vec<String>,
-        timeframe: usize,
-        scrape_url: String,
-        file: Option<PathBuf>,
-    ) -> Self {
-        let data = metrics.iter().map(|m| (m.clone(), vec![])).collect();
-        let data_y_range = metrics.iter().map(|m| (m.clone(), (0.0, 0.0))).collect();
-        let mut file_data: Vec<String> = file
-            .map(|f| std::fs::read_to_string(f).unwrap())
-            .unwrap_or_default()
-            .split('\n')
-            .map(|s| s.to_string())
-            .collect();
-        let mut file_header = vec![];
-        let mut timeframe = timeframe;
-        if !file_data.is_empty() {
-            file_header = file_data[0].split(',').map(|s| s.to_string()).collect();
-            file_data.remove(0);
-
-            while file_data.last().unwrap().is_empty() {
-                file_data.pop();
-            }
-
-            let first_line: Vec<String> = file_data[0].split(',').map(|s| s.to_string()).collect();
-            let last_line: Vec<String> = file_data
-                .last()
-                .unwrap()
-                .split(',')
-                .map(|s| s.to_string())
-                .collect();
-
-            let start_time: usize = first_line.first().unwrap().parse().unwrap();
-            let end_time: usize = last_line.first().unwrap().parse().unwrap();
-
-            timeframe = (end_time - start_time) / 1000;
-        }
-        timeframe = timeframe.clamp(30, 90);
-
-        file_data.reverse();
-        Self {
-            should_quit: false,
-            metrics,
-            start_ts: Instant::now(),
-            data,
-            data_y_range,
-            timeframe,
-            rng: rand::thread_rng(),
-            freeze: false,
-            internal_ts: Duration::default(),
-            scrape_url,
-            file_data,
-            file_header,
-        }
-    }
-
-    /// Chooses what to do when a key is pressed.
-    fn on_key(&mut self, c: char) {
-        match c {
-            'q' => {
-                self.should_quit = true;
-            }
-            'f' => {
-                self.freeze = !self.freeze;
-            }
-            _ => {}
-        }
-    }
-
-    /// Chooses what to do on a tick.
-    async fn on_tick(&mut self) {
-        if self.freeze {
-            return;
-        }
-
-        let metrics_response = match self.file_data.is_empty() {
-            true => {
-                let req = reqwest::Client::new().get(&self.scrape_url).send().await;
-                if req.is_err() {
-                    return;
-                }
-                let data = req.unwrap().text().await.unwrap();
-                iroh_metrics::parse_prometheus_metrics(&data)
-            }
-            false => {
-                if self.file_data.len() == 1 {
-                    self.freeze = true;
-                    return;
-                }
-                let data = self.file_data.pop().unwrap();
-                let r = parse_csv_metrics(&self.file_header, &data);
-                if let Ok(mr) = r {
-                    mr
-                } else {
-                    warn!("Failed to parse csv metrics: {:?}", r.err());
-                    HashMap::new()
-                }
-            }
-        };
-        self.internal_ts = self.start_ts.elapsed();
-        for metric in &self.metrics {
-            let val = if metric.eq("random") {
-                self.rng.gen_range(0..101) as f64
-            } else if let Some(v) = metrics_response.get(metric) {
-                *v
-            } else {
-                0.0
-            };
-            let e = self.data.entry(metric.clone()).or_default();
-            let mut ts = self.internal_ts.as_secs_f64();
-            if metrics_response.contains_key("time") {
-                ts = *metrics_response.get("time").unwrap() / 1000.0;
-            }
-            self.internal_ts = Duration::from_secs_f64(ts);
-            e.push((ts, val));
-            let yr = self.data_y_range.get_mut(metric).unwrap();
-            if val * 1.1 < yr.0 {
-                yr.0 = val * 1.2;
-            }
-            if val * 1.1 > yr.1 {
-                yr.1 = val * 1.2;
-            }
-        }
-    }
-}
-
-/// Parses CSV metrics into a [`HashMap`] of `String` -> `f64`.
-fn parse_csv_metrics(header: &[String], data: &str) -> anyhow::Result<HashMap<String, f64>> {
-    let mut metrics = HashMap::new();
-    let data = data.split(',').collect::<Vec<&str>>();
-    for (i, h) in header.iter().enumerate() {
-        let val = match h.as_str() {
-            "time" => {
-                let ts = data[i].parse::<u64>()?;
-                ts as f64
-            }
-            _ => data[i].parse::<f64>()?,
-        };
-        metrics.insert(h.clone(), val);
-    }
-    Ok(metrics)
 }
 
 /// Deserialise a SecretKey from OpenSSH format.
