+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod commands;
 pub mod config;
 pub mod doctor;
 pub mod logging;
